refactor and start work on parsing & sending
- move code out of main and into utils
- start working on the goroutine to handle packets
